Allow names instead of ids for attributes + cleanup
I did not adapt the scripts yet since we need some special handling for
the attributes which are required by the server directly. So you still
have to use the ids for those. I will change that later.

In the future I want to use the AttributeInfo class instead of the int
id everywhere possible. So I did a small start on that too.

+++ src/game-server/character.h
@@ -324,14 +324,14 @@ class CharacterComponent : public Component
          * basic attribute
          */
         AttribmodResponseCode useCharacterPoint(Entity &entity,
-                                                size_t attribute);
+                                                int attribute);
 
         /**
          * Tries to use a correction point to reduce a
          * basic attribute and regain a character point
          */
         AttribmodResponseCode useCorrectionPoint(Entity &entity,
-                                                 size_t attribute);
+                                                 int attribute);
 
         void setCharacterPoints(int points) { mCharacterPoints = points; }
         int getCharacterPoints() const { return mCharacterPoints; }
